Add new 'get-headers' function
I need to be able to look through all `Received' headers to tell if
something is from bibop, because a crapload of spam is coming in from
there recently.

I decided not to mess with get-header (adding a keyword) because it's
a macro and this really needs to be a function.  If I changed
get-header to a function, it would break email for people.  :(