2.0: new message-in-conversation-p
Support for moving coversations.  Here's the way it works:
a new function in emacs, akin to mh-thread-refile, moves a thread to a
folder, but it also notes information about the conversation (subject,
bhid, message-id's) which incfilter can use to prefer a destination
folder for an incoming message.  This is how it's used in
.mailfilter.cl:

	(let ((dest-folder (message-in-conversation-p subject bhid)))
	  (when dest-folder (return folder)))

Conversations are also an excellent companion to the "mailstatus -l"
command.